requirements.txt
alembic==0.9.1
appdirs==1.4.3
click==6.7
Flask==1.0.2
Flask-Migrate==2.0.3
Flask-Script==2.0.5
Flask-SQLAlchemy==2.3.2
itsdangerous==0.24
Jinja2==2.10
line-bot-sdk==1.8.0
Mako==1.0.6
MarkupSafe==1.0
packaging==16.8
psycopg2==2.7.1
pyparsing==2.2.0
python-editor==1.0.3
six==1.10.0
SQLAlchemy==1.1.9
Werkzeug==0.14.1
azure-common==1.1.16
azure-nspkg==3.0.2
azure-storage==0.36.0
Pillow==5.2.0
requests==2.20.1
pandas==0.23.4